+++ src/phone-input/country-select.js
@@ -7,61 +7,86 @@ LICENSE file in the root directory of this source tree.
 // @flow
 
 import React from 'react';
-import {List, AutoSizer} from 'react-virtualized';
 
 import Flag from './flag.js';
-import {SIZE} from './constants.js';
-import {
-  StyledCountrySelectContainer as Container,
-  StyledCountrySelectListItem as ListItem,
-  StyledCountrySelectFlagContainer as FlagContainer,
-  StyledCountrySelectNameContainer as NameContainer,
-  StyledCountrySelectIsoContainer as IsoContainer,
-} from './styled-components.js';
+import {SIZE, countries} from './constants.js';
+import CountrySelectDropdown from './country-select-dropdown.js';
+import {Select as DefaultSelect} from '../select/index.js';
+import {getOverrides, mergeOverrides} from '../helpers/overrides.js';
 
 import type {CountrySelectPropsT} from './types.js';
 
 export default function CountrySelect(props: CountrySelectPropsT) {
-  const {children, dropdownHeight = '400px', mapIsoToLabel = null} = props;
+  const {
+    countryValue,
+    inputRef,
+    onCountryChange = event => {},
+    size = SIZE.default,
+    dropdownWidth = '400px',
+    dropdownHeight = '400px',
+    mapIsoToLabel,
+    overrides = {
+      CountrySelect: {},
+    },
+  } = props;
+  const baseOverrides = {
+    ValueContainer: {
+      style: {
+        width: {
+          [SIZE.compact]: '34px',
+          [SIZE.default]: '42px',
+          [SIZE.large]: '50px',
+        }[size],
+      },
+    },
+    IconsContainer: {
+      style: {
+        paddingRight: '0',
+      },
+    },
+    SingleValue: {
+      style: {
+        display: 'flex',
+        alignItems: 'center',
+      },
+    },
+    DropdownContainer: {
+      style: {
+        width: dropdownWidth,
+        maxWidth: 'calc(100vw - 10px)',
+      },
+    },
+    Dropdown: {
+      component: CountrySelectDropdown,
+      props: {
+        dropdownHeight: dropdownHeight,
+        mapIsoToLabel: mapIsoToLabel,
+      },
+    },
+  };
+  const [Select, selectProps] = getOverrides(
+    overrides.CountrySelect,
+    DefaultSelect,
+  );
+  const selectOverrides = mergeOverrides(baseOverrides, props.overrides);
   return (
-    <Container $height={dropdownHeight}>
-      <AutoSizer>
-        {({height, width}) => {
-          return (
-            <List
-              role="listbox"
-              height={height}
-              width={width}
-              rowCount={children.length}
-              rowHeight={32}
-              rowRenderer={({index, key, style}) => {
-                // resetMenu and getItemLabel should not end up on native html elements
-                const {resetMenu, getItemLabel, ...rest} = children[
-                  index
-                ].props;
-                return (
-                  <ListItem key={key} style={style} {...rest}>
-                    <FlagContainer>
-                      <Flag
-                        iso2={children[index].props.item.id}
-                        size={SIZE.compact}
-                      />
-                    </FlagContainer>
-                    <NameContainer>
-                      {mapIsoToLabel
-                        ? mapIsoToLabel(props.children[index].props.item.id)
-                        : children[index].props.item.label}
-                    </NameContainer>
-                    <IsoContainer>
-                      +{children[index].props.item.dialCode}
-                    </IsoContainer>
-                  </ListItem>
-                );
-              }}
-            />
-          );
-        }}
-      </AutoSizer>
-    </Container>
+    <Select
+      size={size}
+      value={[countryValue]}
+      onChange={event => {
+        if (inputRef && inputRef.current) {
+          inputRef.current.focus();
+        }
+        onCountryChange(event);
+      }}
+      options={countries}
+      clearable={false}
+      searchable={false}
+      getValueLabel={({option}) => {
+        return option.id ? <Flag iso2={String(option.id)} size={size} /> : null;
+      }}
+      overrides={selectOverrides}
+      {...selectProps}
+    />
   );
 }
