+++ src/phone-input/phone-input.js
@@ -9,10 +9,8 @@ LICENSE file in the root directory of this source tree.
 import React, {useRef} from 'react';
 
 import {Input as DefaultInput} from '../input/index.js';
-import {countries, SIZE} from './constants.js';
+import {SIZE} from './constants.js';
 import CountrySelect from './country-select.js';
-import Flag from './flag.js';
-import {Select as DefaultSelect} from '../select/index.js';
 import {getOverrides, mergeOverrides} from '../helpers/overrides.js';
 
 import type {PropsT} from './types.js';
@@ -20,16 +18,11 @@ import type {PropsT} from './types.js';
 export default function PhoneInput(props: PropsT) {
   const {
     inputValue,
-    countryValue,
     onInputChange,
-    onCountryChange,
     size = SIZE.default,
-    dropdownHeight = '400px',
-    dropdownWidth = '400px',
-    mapIsoToLabel = null,
     overrides = {
       Input: {},
-      Select: {},
+      CountrySelect: {},
     },
   } = props;
   const inputRef = useRef(null);
@@ -40,69 +33,13 @@ export default function PhoneInput(props: PropsT) {
       },
     },
     Before: {
-      component: function Before(props) {
-        const baseOverrides = {
-          ValueContainer: {
-            style: {
-              width: {
-                [SIZE.compact]: '34px',
-                [SIZE.default]: '42px',
-                [SIZE.large]: '50px',
-              }[size],
-            },
-          },
-          IconsContainer: {
-            style: {
-              paddingRight: '0',
-            },
-          },
-          SingleValue: {
-            style: {
-              display: 'flex',
-              alignItems: 'center',
-            },
-          },
-          DropdownContainer: {
-            style: {
-              width: dropdownWidth,
-              maxWidth: 'calc(100vw - 10px)',
-            },
-          },
-          Dropdown: {
-            component: CountrySelect,
-            props: {
-              dropdownHeight: dropdownHeight,
-              mapIsoToLabel: mapIsoToLabel,
-            },
-          },
-        };
-        const [Select, selectProps] = getOverrides(
-          overrides.Select,
-          DefaultSelect,
-        );
-        const selectOverrides = mergeOverrides(baseOverrides, props.overrides);
-        return (
-          <Select
-            size={size}
-            value={[countryValue]}
-            onChange={event => {
-              if (inputRef && inputRef.current) {
-                inputRef.current.focus();
-              }
-              onCountryChange(event);
-            }}
-            options={countries}
-            clearable={false}
-            searchable={false}
-            getValueLabel={({option}) => {
-              return option.id ? (
-                <Flag iso2={String(option.id)} size={size} />
-              ) : null;
-            }}
-            overrides={selectOverrides}
-            {...selectProps}
-          />
-        );
+      component: CountrySelect,
+      props: {
+        ...props,
+        inputRef,
+        overrides: {
+          CountrySelect: overrides.CountrySelect,
+        },
       },
     },
   };
